Consolidate logic for showing/hiding toast and removing those temp proposals from localstorage
